Add tag-release.sh script for bumping version & tagging
I put all into a shell script so you could invoke it from anywhere if you decide to further automate the release (locally or on the CI).

The script takes the new version as an argument and does 4 things:
1. Check that there are no uncommitted changes (otherwise fails).
2. Replace the version String in the SDK.
3. Commit the changes.
4. Tag the new commit.

I decided against a lint check because such a check would require the tag to be already available but you created the latest tag _after_ merging https://github.com/TelemetryDeck/SwiftClient/pull/70, so a lint check would run when it's already too late. The same problem exists with pre-commit hooks, as a tag can only be created _after_ the commit is made.

Fixes #57.