Ensure not to make SwiftUI previews crash due to missing init call
I found and used this method for checking if we're running for SwiftUI previews [here](https://stackoverflow.com/a/61741858/3451975).
I made sure to order the `if` checks so and cache the result in case of sending signals, that the performance when running in a real world application shouldn't get affected.

Note that I followed the exact same implementation as for `testMode` to ensure users can always override the behavior by setting a different value for `swiftUIPreviewMode`, e.g. in case they actually run the application from SwiftUI for whatever reason instead of using the Run button in Xcode.

I'm not sure if `swiftUIPreviewMode` is the best name though. It describes the default behavior very well, but I can imagine someone might use this same setting and set it to `true` for turning off analytics when a user disables such an option in an apps settings (opt-out), therefore maybe the property could be renamed to something like `sendSignals`. But that's a change separate from fixing just the SwiftUI issue, so I was not sure if you want to introduce such an option.

Fixes https://github.com/TelemetryDeck/SwiftClient/issues/73.