send extension identifier, if provided
iOS and other Apple operating systems have a plethora of extension types all with a unique identifier which can be retrieved from the Info.plist for the current target.

A complete list of identifiers can be found here: https://developer.apple.com/documentation/bundleresources/information_property_list/nsextension/nsextensionpointidentifier

It is useful for developers to understand the source of the events, including whether it came from a widget, notification service, or other extension.