Do not block calling queue with IO operations of `SignalCache`
`SignalCache` used to dispatch synchronously instead of asynchronously to its concurrent queue for the IO operations of loading the cache and backing up the cache to file system which are typically done on app launch and when the app is sent to background respectively. Given how these functions are called, these would be typically blocking the main thread.