Add new method TelemetryClient.requestImmediateSync() for advanced churn optimization
This new method should only be used in rare cases where an important information is sent to the servers that must reach the servers even if the user might leave the app right after and never return.
This can be useful for apps that are used only once to track important last steps of the user. Or for experiences like errors that could make the user delete the app right after.
Those error signals would never reach the server without this.