Remove the old “newSessionBegan” signal
We’re usually sending old-style parameter names in parallel with the new-style parameter so that people who switch from SDK version 1 to 2 have an easier and smoother upgrade path.

However, this leads to problems with customers with lots of users, because they get twice as many session signals now.

Since not too many customers rely on session signals right now, I’d suggest we just do a hard switch in this case and deal with the slightly increased support load by offering customers who write in abou this updated queries.